Add DynamicallyComposedMultiHeadAttention class in attention.py
A new class called DynamicallyComposedMultiHeadAttention has been added to the attention.py file in the composed_detr project. This class includes initialization parameters, various helper functions, and a forward function for implementing the computation of attention. It also includes a sample usage of this class at the end of the file.